+++ CHANGELOG.md
@@ -6,6 +6,11 @@ For more information about changelogs, check
 [Keep a Changelog](http://keepachangelog.com) and
 [Vandamme](http://tech-angels.github.io/vandamme).
 
+## 0.25.27 - 2016-03-28
+
+* [FEATURE] Add `comment_threads` association to Yt::Video.
+* [FEATURE] Add `top_level_comment` and delegate its attributes (`text_display`, `author_display_name`, `like_count`, `updated_at`) to Yt::CommentThread.
+
 ## 0.25.26 - 2016-03-24
 
 * [FEATURE] Add Yt::Comment resource.